requirements.txt
fastapi;
inngest;
llama-index-core;
llama-index-readers-file;
python-dotenv;
qdrant-client;
uvicorn[standard];
streamlit;
